Move ``_get_render_mode`` to ``_shared.py``
