+++ sphinx/ext/autodoc/_generate.py
@@ -9,6 +9,7 @@ from sphinx.errors import PycodeError
 from sphinx.ext.autodoc._member_finder import _gather_members
 from sphinx.ext.autodoc._renderer import _add_content, _directive_header_lines
 from sphinx.ext.autodoc._sentinels import ALL
+from sphinx.ext.autodoc._shared import _get_render_mode
 from sphinx.ext.autodoc.mock import ismock
 from sphinx.locale import _, __
 from sphinx.pycode import ModuleAnalyzer
@@ -17,7 +18,6 @@ from sphinx.util.typing import restify, stringify_annotation
 
 if TYPE_CHECKING:
     from collections.abc import Iterator, Mapping, MutableSet
-    from typing import Literal
 
     from sphinx.config import Config
     from sphinx.environment import _CurrentDocument
@@ -179,8 +179,9 @@ def _add_directive_lines(
     docstring_lines = StringList(props.docstring_lines, source=source_name)
 
     # add alias information, if applicable
+    render_mode = _get_render_mode(config.autodoc_typehints_format)
     lines = _body_alias_lines(
-        typehints_format=config.autodoc_typehints_format,
+        render_mode=render_mode,
         python_display_short_literal_types=config.python_display_short_literal_types,
         props=props,
     )
@@ -278,7 +279,7 @@ def _document_members(
 def _body_alias_lines(
     *,
     props: _ItemProperties,
-    typehints_format: Literal['fully-qualified', 'short'],
+    render_mode: _RestifyMode,
     python_display_short_literal_types: bool,
 ) -> Iterator[str]:
     """Add content from docstrings, attribute documentation and user."""
@@ -287,11 +288,9 @@ def _body_alias_lines(
 
         assert isinstance(props, _AssignStatementProperties)
 
-        mode = _get_render_mode(typehints_format)
-
         # Support for documenting GenericAliases
         if props._obj_is_generic_alias:
-            alias = restify(props._obj, mode=mode)
+            alias = restify(props._obj, mode=render_mode)
             yield _('alias of %s') % alias
             yield ''
             return
@@ -303,10 +302,9 @@ def _body_alias_lines(
         assert isinstance(props, _ClassDefProperties)
 
         obj = props._obj
-        mode = _get_render_mode(typehints_format)
 
         if props._obj_is_new_type:
-            supertype = restify(obj.__supertype__, mode=mode)
+            supertype = restify(obj.__supertype__, mode=render_mode)
             yield _('alias of %s') % supertype
             yield ''
             return
@@ -317,13 +315,13 @@ def _body_alias_lines(
                 repr(obj.__name__),
                 *(
                     stringify_annotation(
-                        constraint, mode, short_literals=short_literals
+                        constraint, render_mode, short_literals=short_literals
                     )
                     for constraint in obj.__constraints__
                 ),
             ]
             if obj.__bound__:
-                attrs.append(rf'bound=\ {restify(obj.__bound__, mode=mode)}')
+                attrs.append(rf'bound=\ {restify(obj.__bound__, mode=render_mode)}')
             if obj.__covariant__:
                 attrs.append('covariant=True')
             if obj.__contravariant__:
@@ -345,7 +343,7 @@ def _body_alias_lines(
 
             if class_var_doc_comment:
                 return
-            alias = restify(obj, mode=mode)
+            alias = restify(obj, mode=render_mode)
             yield _('alias of %s') % alias
             return
 
@@ -367,11 +365,3 @@ def _docstring_source_name(*, props: _ItemProperties, source: str) -> str:
     if source:
         return f'{source}:docstring of {fullname}'
     return f'docstring of {fullname}'
-
-
-def _get_render_mode(
-    typehints_format: Literal['fully-qualified', 'short'],
-) -> _RestifyMode:
-    if typehints_format == 'short':
-        return 'smart'
-    return 'fully-qualified-except-typing'
